Add new file Styling struct
